add flag to prevent doing redis server `flushall()`
 * prevent non-test instances from being wiped clean

+++ tests/conftest.py
@@ -15,7 +15,8 @@ def d():
         host=os.environ.get("TEST_REDIS_HOST", "localhost"),
         port=os.environ.get("TEST_REDIS_PORT", 63798),
     )  # use a different port than usual because are clearing it!
-    redis_client.flushall()
+    if os.environ.get("CLEAR_REDIS", "nottrue") == "true":
+        redis_client.flushall()
     prefix = uuid.uuid4().hex
     yield RedisJSONDict(redis_client, prefix=prefix)
     # Clean up.
